User commit

+++ src/main/java/com/example/demo/model/Recommendation.java
@@ -2,6 +2,7 @@ package com.example.demo.model;
 
 import jakarta.persistence.*;
 import lombok.*;
+import java.math.BigDecimal;
 
 import java.time.LocalDateTime;
 
@@ -26,7 +27,7 @@ public class Recommendation {
 
     private String basisSnapshot;
 
-    private double confidenceScore;
+    private BigDecimal confidenceScore;
 
     @PrePersist
     public void prePersist() {
